Created a 'controllers' folder for the /signin, /register, and /profile routes to clean up the server.js file
