+++ controllers/profile.js
@@ -1,0 +1,16 @@
+const handleProfileGet = (req, res, db) => {
+	const { id } = req.params;
+	db.select('*').from('users').where({id})
+		.then(user => {
+		if (user.length) {
+			res.json(user[0])
+		} else {
+			res.status(400).json('Not Found')
+		}
+	})
+	.catch(err => res.status(400).json('Not Found'))
+}
+
+module.exports = {
+	handleProfileGet
+}
